refactor: MineStrategy 인터페이스, 패키지 네이밍 변경
