+++ src/main/kotlin/minesweeper/domain/board/strategy/MineStrategy.kt
@@ -1,0 +1,8 @@
+package minesweeper.domain.board.strategy
+
+import minesweeper.domain.common.PositiveInt
+
+interface MineStrategy {
+
+    fun strategy(): (numberOfCells: PositiveInt, numberOfMines: PositiveInt) -> List<Int>
+}
