+++ src/main/kotlin/minesweeper/game/MineSweeper.kt
@@ -1,12 +1,12 @@
 package minesweeper.game
 
-import minesweeper.domain.board.random.DefaultRandomMineStrategy
+import minesweeper.domain.board.strategy.DefaultRandomMineStrategy
 import minesweeper.view.ViewResolver
 
 fun main() {
     val mineSweeperController = MineSweeperController(
         viewResolver = ViewResolver(),
-        randomMineStrategy = DefaultRandomMineStrategy()
+        mineStrategy = DefaultRandomMineStrategy()
     )
     mineSweeperController.start()
 }
