+++ src/main/kotlin/minesweeper/game/MineSweeperController.kt
@@ -1,12 +1,12 @@
 package minesweeper.game
 
 import minesweeper.domain.board.mineBoard
-import minesweeper.domain.board.random.RandomMineStrategy
+import minesweeper.domain.board.strategy.MineStrategy
 import minesweeper.view.ViewResolver
 
 class MineSweeperController(
     private val viewResolver: ViewResolver,
-    private val randomMineStrategy: RandomMineStrategy
+    private val mineStrategy: MineStrategy
 ) {
 
     fun start() {
@@ -15,7 +15,7 @@ class MineSweeperController(
             width(request.width)
             height(request.height)
             numberOfMines(request.numberOfMines)
-            mineStrategy(randomMineStrategy.strategy())
+            mineStrategy(mineStrategy.strategy())
         }
         viewResolver.printStartOfGame(mineBoard)
     }
